chore: Refactor conversions from dao to be static methods on models
