+++ app/models/storage/conversions.from-dao.ts
@@ -1,38 +1,7 @@
 import { google, LiftLog } from '@/gen/proto';
-import {
-  SessionBlueprint,
-  WeightedExerciseBlueprint,
-  Rest,
-  ProgramBlueprint,
-  ExerciseBlueprint,
-  CardioExerciseBlueprint,
-  CardioTarget,
-  WeightedExerciseBlueprintPOJO,
-  CardioExerciseBlueprintPOJO,
-  DistanceUnit,
-  CardioExerciseSetBlueprint,
-} from '@/models/blueprint-models';
-import {
-  PotentialSet,
-  RecordedWeightedExercise,
-  RecordedSet,
-  Session,
-  RecordedExercise,
-  RecordedCardioExercise,
-  RecordedCardioExerciseSet,
-} from '@/models/session-models';
+import { uuidStringify } from '@/utils/uuid';
 import Long from 'long';
-import {
-  FeedIdentity,
-  FeedItem,
-  FeedUser,
-  FollowRequest,
-  FollowRequestPOJO,
-  SessionFeedItem,
-  SharedItem,
-  SharedProgramBlueprint,
-  SharedSession,
-} from '@/models/feed-models';
+import { UuidConversionError } from './uuid-conversion-error';
 import {
   Duration,
   Instant,
@@ -42,14 +11,7 @@ import {
   ZoneOffset,
 } from '@js-joda/core';
 import BigNumber from 'bignumber.js';
-import { UuidConversionError } from '@/models/storage/uuid-conversion-error';
-import { FeedState } from '@/store/feed';
-import { RemoteData } from '@/models/remote';
-import { uuidStringify } from '@/utils/uuid';
-import { Weight, WeightUnit } from '@/models/weight';
-import { match, P } from 'ts-pattern';
 
-// Converts a UUID DAO to a string
 export function fromUuidDao(
   dao: LiftLog.Ui.Models.IUuidDao | null | undefined,
 ): string {
@@ -144,332 +106,6 @@ export function fromTimestampDao(
   return Instant.ofEpochSecond(dao!.seconds!.toNumber());
 }
 
-// Converts a RecordedSet DAO to a RecordedSetPOJO
-export function fromRecordedSetDao(
-  sessionDate: LiftLog.Ui.Models.IDateOnlyDao,
-  recordedSetDao: LiftLog.Ui.Models.SessionHistoryDao.IRecordedSetDaoV2,
-): RecordedSet {
-  const dateCompleted = recordedSetDao.completionDate ?? sessionDate;
-  const completionLocalDateTime = fromDateOnlyDao(dateCompleted).atTime(
-    fromTimeOnlyDao(recordedSetDao.completionTime),
-  );
-  const completionDateTime = completionLocalDateTime.atOffset(
-    recordedSetDao.completionOffset
-      ? ZoneOffset.ofTotalSeconds(recordedSetDao.completionOffset.totalSeconds!)
-      : ZoneOffset.systemDefault()
-          .rules()
-          .offsetOfLocalDateTime(completionLocalDateTime),
-  );
-
-  return RecordedSet.fromPOJO({
-    completionDateTime,
-    repsCompleted: recordedSetDao.repsCompleted!,
-  });
-}
-
-// Converts a PotentialSet DAO to a PotentialSetPOJO
-function fromPotentialSetDao(
-  sessionDate: LiftLog.Ui.Models.IDateOnlyDao,
-  dao:
-    | LiftLog.Ui.Models.SessionHistoryDao.IPotentialSetDaoV2
-    | null
-    | undefined,
-): PotentialSet {
-  if (!dao) {
-    throw new Error('PotentialSetDao cannot be null');
-  }
-  return PotentialSet.fromPOJO({
-    set: dao.recordedSet
-      ? fromRecordedSetDao(sessionDate, dao.recordedSet).toPOJO()
-      : undefined,
-    weight: fromWeight(
-      fromDecimalDao(dao.weightValue) ?? BigNumber(0),
-      dao.weightUnit,
-    ),
-  });
-}
-
-// Converts a RecordedExercise DAO to a RecordedExercisePOJO
-export function fromRecordedExerciseDao(
-  sessionDate: LiftLog.Ui.Models.IDateOnlyDao,
-  dao:
-    | LiftLog.Ui.Models.SessionHistoryDao.IRecordedExerciseDaoV2
-    | null
-    | undefined,
-): RecordedExercise {
-  if (!dao) {
-    throw new Error('Recorded exercise DAO cannot be null');
-  }
-  if (dao.type === LiftLog.Ui.Models.SessionBlueprintDao.ExerciseType.CARDIO) {
-    const sets = dao.cardioSets!.map((x) => fromCardioSetDao(x).toPOJO());
-    return RecordedCardioExercise.fromPOJO({
-      notes: dao.notes?.value ?? undefined,
-      blueprint: fromExerciseBlueprintDao(
-        dao.exerciseBlueprint,
-      ).toPOJO() as CardioExerciseBlueprintPOJO,
-      sets:
-        sets.length === 0
-          ? [getRecordedCardioSetFromDeprecatedFields(dao).toPOJO()]
-          : sets,
-    });
-  }
-  return RecordedWeightedExercise.fromPOJO({
-    notes: dao.notes?.value ?? undefined,
-    blueprint: fromExerciseBlueprintDao(
-      dao.exerciseBlueprint,
-    ).toPOJO() as WeightedExerciseBlueprintPOJO,
-    potentialSets: dao.potentialSets!.map((x) =>
-      fromPotentialSetDao(sessionDate, x).toPOJO(),
-    ),
-  });
-}
-
-function fromCardioSetDao(
-  dao: LiftLog.Ui.Models.SessionHistoryDao.IRecordedCardioExerciseSetDao,
-): RecordedCardioExerciseSet {
-  return RecordedCardioExerciseSet.fromPOJO({
-    blueprint: fromCardioExerciseSetBlueprint(dao.blueprint!).toPOJO(),
-    distance:
-      dao.distanceValue && dao.distanceUnit
-        ? {
-            value: fromDecimalDao(dao.distanceValue),
-            unit: dao.distanceUnit.value as DistanceUnit,
-          }
-        : undefined,
-    duration: fromDurationDao(dao.duration),
-    completionDateTime: fromDateTimeDao(dao.completionDateTime),
-    incline: fromDecimalDao(dao.incline),
-    resistance: fromDecimalDao(dao.resistance),
-    weight: dao.weight ? fromWeightDao(dao.weight) : undefined,
-    steps: dao.steps?.value ?? undefined,
-    currentBlockStartTime: undefined,
-  });
-}
-
-function getRecordedCardioSetFromDeprecatedFields(
-  dao: LiftLog.Ui.Models.SessionHistoryDao.IRecordedExerciseDaoV2,
-): RecordedCardioExerciseSet {
-  return RecordedCardioExerciseSet.fromPOJO({
-    blueprint: getCardioBlueprintSetFromDeprecatedFields(
-      dao.exerciseBlueprint!,
-    ).toPOJO(),
-    distance:
-      dao.deprecatedDistanceValue && dao.deprecatedDistanceUnit
-        ? {
-            value: fromDecimalDao(dao.deprecatedDistanceValue),
-            unit: dao.deprecatedDistanceUnit.value as DistanceUnit,
-          }
-        : undefined,
-    duration: fromDurationDao(dao.deprecatedDuration),
-    completionDateTime: fromDateTimeDao(dao.deprecatedCompletionDateTime),
-    incline: fromDecimalDao(dao.deprecatedIncline),
-    resistance: fromDecimalDao(dao.deprecatedResistance),
-    weight: undefined,
-    steps: undefined,
-    currentBlockStartTime: undefined,
-  });
-}
-
-// Converts a Session DAO to a SessionPOJO
-export function fromSessionDao(
-  dao: LiftLog.Ui.Models.SessionHistoryDao.ISessionDaoV2 | null | undefined,
-): Session {
-  if (!dao) {
-    throw new Error('Session dao cannot be null');
-  }
-  const recordedExercises =
-    dao.recordedExercises?.map((x) =>
-      fromRecordedExerciseDao(dao.date!, x).toPOJO(),
-    ) ?? [];
-  return Session.fromPOJO({
-    id: fromUuidDao(dao.id),
-    blueprint: SessionBlueprint.fromPOJO({
-      name: dao.sessionName!,
-      exercises: recordedExercises.map((x) => x.blueprint),
-      notes: dao.blueprintNotes ?? '',
-    }).toPOJO(),
-    bodyweight: dao.bodyweightValue
-      ? fromWeight(fromDecimalDao(dao.bodyweightValue), dao.bodyweightUnit)
-      : undefined,
-    date: fromDateOnlyDao(dao.date),
-    recordedExercises,
-  });
-}
-
-function fromWeightUnitDao(
-  daoUnit: LiftLog.Ui.Models.WeightUnit | null | undefined,
-): WeightUnit {
-  return match(daoUnit)
-    .returnType<WeightUnit>()
-    .with(P.nullish, () => 'nil')
-    .with(LiftLog.Ui.Models.WeightUnit.NIL satisfies 0 as 0, () => 'nil')
-    .with(
-      LiftLog.Ui.Models.WeightUnit.KILOGRAMS satisfies 1 as 1,
-      () => 'kilograms',
-    )
-    .with(LiftLog.Ui.Models.WeightUnit.POUNDS satisfies 2 as 2, () => 'pounds')
-    .exhaustive();
-}
-
-function fromWeight(
-  value: BigNumber,
-  daoUnit: LiftLog.Ui.Models.WeightUnit | null | undefined,
-): Weight {
-  return new Weight(value, fromWeightUnitDao(daoUnit));
-}
-
-// Converts a SessionHistory DAO to a Map of Sessions
-export function fromSessionHistoryDao(
-  sessionHistoryModel: LiftLog.Ui.Models.SessionHistoryDao.SessionHistoryDaoV2,
-): Map<string, Session> {
-  return sessionHistoryModel.completedSessions.reduce((map, item) => {
-    map.set(
-      fromUuidDao(item.id),
-      Session.fromPOJO(fromSessionDao(item).toPOJO()),
-    );
-    return map;
-  }, new Map<string, Session>());
-}
-
-// Converts a SessionBlueprint DAO to a SessionBlueprint
-export function fromSessionBlueprintDao(
-  dao: LiftLog.Ui.Models.SessionBlueprintDao.ISessionBlueprintDaoV2,
-): SessionBlueprint {
-  return new SessionBlueprint(
-    dao.name!,
-    dao.exerciseBlueprints!.map(fromExerciseBlueprintDao),
-    dao.notes ?? '',
-  );
-}
-
-export function fromExerciseBlueprintDao(
-  dao:
-    | LiftLog.Ui.Models.SessionBlueprintDao.IExerciseBlueprintDaoV2
-    | null
-    | undefined,
-): ExerciseBlueprint {
-  if (!dao) {
-    throw new Error('ExerciseBlueprint dao should not be null');
-  }
-  if (dao.type === LiftLog.Ui.Models.SessionBlueprintDao.ExerciseType.CARDIO) {
-    const sets = dao.cardioSets!.map(fromCardioExerciseSetBlueprint);
-    return new CardioExerciseBlueprint(
-      dao.name!,
-      sets.length === 0
-        ? [getCardioBlueprintSetFromDeprecatedFields(dao)]
-        : sets,
-      dao.notes ?? '',
-      dao.link ?? '',
-    );
-  }
-  return new WeightedExerciseBlueprint(
-    dao.name!,
-    dao.sets!,
-    dao.repsPerSet!,
-    fromDecimalDao(dao.weightIncreaseOnSuccess) ?? BigNumber(0),
-    fromRestDao(dao.restBetweenSets),
-    dao.supersetWithNext ?? false,
-    dao.notes ?? '',
-    dao.link ?? '',
-  );
-}
-
-function fromCardioExerciseSetBlueprint(
-  dao: LiftLog.Ui.Models.SessionBlueprintDao.ICardioExerciseSetBlueprintDao,
-): CardioExerciseSetBlueprint {
-  return CardioExerciseSetBlueprint.fromPOJO({
-    target: fromCardioTargetDao(dao.cardioTarget),
-    trackDuration: dao.trackDuration ?? false,
-    trackDistance: dao.trackDistance ?? false,
-    trackResistance: dao.trackResistance ?? false,
-    trackIncline: dao.trackIncline ?? false,
-    trackWeight: dao.trackWeight ?? false,
-    trackSteps: dao.trackSteps ?? false,
-  });
-}
-
-function getCardioBlueprintSetFromDeprecatedFields(
-  dao: LiftLog.Ui.Models.SessionBlueprintDao.IExerciseBlueprintDaoV2,
-): CardioExerciseSetBlueprint {
-  return CardioExerciseSetBlueprint.fromPOJO({
-    target: fromCardioTargetDao(dao.deprecatedCardioTarget),
-    trackDuration: dao.deprecatedTrackDuration ?? false,
-    trackDistance: dao.deprecatedTrackDistance ?? false,
-    trackResistance: dao.deprecatedTrackResistance ?? false,
-    trackIncline: dao.deprecatedTrackIncline ?? false,
-    trackWeight: false,
-    trackSteps: false,
-  });
-}
-
-function fromCardioTargetDao(
-  dao: LiftLog.Ui.Models.SessionBlueprintDao.ICardioTarget | null | undefined,
-): CardioTarget {
-  if (!dao) {
-    throw new Error('Expected a non null cardio target');
-  }
-  return {
-    type: dao.type as CardioTarget['type'],
-    value: match(dao.type as CardioTarget['type'])
-      .with('distance', () => ({
-        value: fromDecimalDao(dao.distanceValue) ?? BigNumber(0),
-        unit: dao.distanceUnit ?? 'metre',
-      }))
-      .with('time', () => fromDurationDao(dao.timeValue))
-      .exhaustive(),
-  } as CardioTarget;
-}
-
-export function fromRestDao(
-  dao: LiftLog.Ui.Models.SessionBlueprintDao.IRestDaoV2 | null | undefined,
-): Rest {
-  return {
-    minRest: fromDurationDao(dao?.minRest) ?? Duration.ZERO,
-    maxRest: fromDurationDao(dao?.maxRest) ?? Duration.ZERO,
-    failureRest: fromDurationDao(dao?.failureRest) ?? Duration.ZERO,
-  };
-}
-
-// Converts a ProgramBlueprint DAO to a ProgramBlueprint
-export function fromProgramBlueprintDao(
-  dao: LiftLog.Ui.Models.ProgramBlueprintDao.IProgramBlueprintDaoV1,
-): ProgramBlueprint {
-  return new ProgramBlueprint(
-    dao.name!,
-    dao.sessions!.map(fromSessionBlueprintDao),
-    dao.lastEdited ? fromDateOnlyDao(dao.lastEdited) : LocalDate.now(),
-  );
-}
-
-export function fromFeedStateDao(dao: LiftLog.Ui.Models.IFeedStateDaoV1) {
-  return {
-    feed: dao.feedItems?.map(fromFeedItemDao).map((x) => x.toPOJO()) ?? [],
-    followedUsers:
-      (dao.followedUsers &&
-        Object.fromEntries(
-          dao.followedUsers
-            .map(fromFeedUserDao)
-            .map((x) => [x.id, x.toPOJO()] as const),
-        )) ??
-      {},
-    identity:
-      (dao.identity &&
-        RemoteData.success(fromFeedIdentityDao(dao.identity).toPOJO())) ??
-      RemoteData.notAsked(),
-    followRequests: dao.followRequests?.map(fromFollowRequestDao) ?? [],
-    followers:
-      (dao.followers &&
-        Object.fromEntries(
-          dao.followers
-            .map(fromFeedUserDao)
-            .map((x) => [x.id, x.toPOJO()] as const),
-        )) ??
-      {},
-    unpublishedSessionIds: dao.unpublishedSessionIds?.map(fromUuidDao) ?? [],
-    revokedFollowSecrets: dao.revokedFollowSecrets ?? [],
-  } satisfies Partial<FeedState>;
-}
-
 export function fromDurationDao(
   duration: google.protobuf.IDuration | null | undefined,
 ) {
@@ -479,118 +115,4 @@ export function fromDurationDao(
   return Duration.ofSeconds(
     Long.fromValue(duration.seconds!).toNumber(),
   ).plusNanos(Long.fromValue(duration.nanos!).toNumber());
-}
-
-export function fromFeedItemDao(
-  dao: LiftLog.Ui.Models.IFeedItemDaoV1,
-): FeedItem {
-  return new SessionFeedItem(
-    fromUuidDao(dao.userId),
-    fromUuidDao(dao.eventId),
-    fromTimestampDao(dao.timestamp),
-    fromTimestampDao(dao.expiry),
-    fromSessionDao(dao.session),
-  );
-}
-
-// Converts a FeedIdentity DAO to a FeedIdentity
-export function fromFeedIdentityDao(
-  dao: LiftLog.Ui.Models.IFeedIdentityDaoV1,
-): FeedIdentity {
-  return new FeedIdentity(
-    fromUuidDao(dao.id),
-    dao.lookup?.value ?? '',
-    { value: Uint8Array.from(dao.aesKey!) },
-    {
-      publicKey: { spkiPublicKeyBytes: Uint8Array.from(dao.publicKey!) },
-      privateKey: { pkcs8PrivateKeyBytes: Uint8Array.from(dao.privateKey!) },
-    },
-    dao.password!,
-    dao.name?.value ?? undefined,
-    (dao.profilePicture && Uint8Array.from(dao.profilePicture)) ?? undefined,
-    dao.publishBodyweight ?? false,
-    dao.publishPlan ?? false,
-    dao.publishWorkouts ?? false,
-  );
-}
-
-// Converts a FeedUser DAO to a FeedUser
-export function fromFeedUserDao(
-  dao: LiftLog.Ui.Models.IFeedUserDaoV1,
-): FeedUser {
-  return new FeedUser(
-    fromUuidDao(dao.id),
-    { spkiPublicKeyBytes: Uint8Array.from(dao.publicKey!) },
-    dao.name?.value ?? undefined,
-    dao.nickname?.value ?? undefined,
-    dao.currentPlan ? fromCurrentPlanDao(dao.currentPlan) : [],
-    (dao.profilePicture && Uint8Array.from(dao.profilePicture)) ?? undefined,
-    dao.aesKey?.length ? { value: Uint8Array.from(dao.aesKey) } : undefined,
-    dao.followSecret?.value ?? undefined,
-  );
-}
-
-// Converts a CurrentPlan DAO to a CurrentPlan
-export function fromCurrentPlanDao(
-  dao: LiftLog.Ui.Models.ICurrentPlanDaoV1,
-): SessionBlueprint[] {
-  return dao.sessions!.map(fromSessionBlueprintDao);
-}
-
-export function fromCurrentSessionDao(
-  dao: LiftLog.Ui.Models.CurrentSessionStateDao.ICurrentSessionStateDaoV2,
-) {
-  return {
-    workoutSession: dao.workoutSession && fromSessionDao(dao.workoutSession),
-    historySession: dao.historySession && fromSessionDao(dao.historySession),
-  };
-}
-
-export function fromSharedItemDao(
-  dao: LiftLog.Ui.Models.SharedItemPayload,
-): SharedItem | null {
-  switch (dao.payload) {
-    case 'sharedProgramBlueprint': {
-      const programBlueprintDao = dao.sharedProgramBlueprint?.programBlueprint;
-      if (!programBlueprintDao) {
-        return null;
-      }
-      return new SharedProgramBlueprint(
-        fromProgramBlueprintDao(programBlueprintDao),
-      );
-    }
-    case 'sharedSession': {
-      const sessionDao = dao.sharedSession?.session;
-      if (!sessionDao) {
-        return null;
-      }
-      return new SharedSession(fromSessionDao(sessionDao));
-    }
-    case undefined:
-      return null;
-    default: {
-      // We don't want it throwing for older clients
-      expectNever(dao.payload);
-      return null;
-    }
-  }
-}
-function fromFollowRequestDao(
-  value: LiftLog.Ui.Models.IInboxMessageDao,
-): FollowRequestPOJO {
-  return FollowRequest.fromPOJO({
-    name: value.followRequest?.name?.value ?? '',
-    userId: fromUuidDao(value.fromUserId),
-  }).toPOJO();
-}
-
-export function fromWeightDao(value: LiftLog.Ui.Models.IWeight): Weight {
-  return new Weight(
-    fromDecimalDao(value.value!),
-    fromWeightUnitDao(value.unit),
-  );
-}
-
-function expectNever(never: never) {
-  // Do nothing, this is a compile time check
 }
