+++ app/models/storage/conversions.spec.ts
@@ -12,13 +12,7 @@ import {
   fromDecimalDao,
   fromDateOnlyDao,
   fromTimeOnlyDao,
-  fromSessionBlueprintDao,
-  fromSessionDao,
-  fromFeedIdentityDao,
-  fromFeedUserDao,
-  fromSessionHistoryDao,
   fromDurationDao,
-  fromFeedItemDao,
   fromTimestampDao,
 } from '@/models/storage/conversions.from-dao';
 import BigNumber from 'bignumber.js';
@@ -41,6 +35,9 @@ import { describe, it, expect } from 'vitest';
 import { readFileSync } from 'fs';
 import { gunzipSync } from 'zlib';
 import { Weight } from '@/models/weight';
+import { SessionBlueprint } from '../blueprint-models';
+import { fromSessionHistoryDao, Session } from '../session-models';
+import { FeedIdentity, FeedUser, SessionFeedItem } from '../feed-models';
 
 const Models = LiftLog.Ui.Models;
 
@@ -52,17 +49,17 @@ const modelToDao = (x: ToDao) => x.toDao();
 
 describe('conversions', () => {
   it.each`
-    name                  | protoType                                           | initialValueGenerator        | convertToDao      | convertFromDao             | assertEquals
-    ${'Decimal'}          | ${Models.DecimalValue}                              | ${BigNumberGenerator}        | ${toDecimalDao}   | ${fromDecimalDao}          | ${(a: BigNumber, b: BigNumber) => expect(a.isEqualTo(b)).toBeTruthy()}
-    ${'DateOnly'}         | ${Models.DateOnlyDao}                               | ${LocalDateGenerator}        | ${toDateOnlyDao}  | ${fromDateOnlyDao}         | ${(a: LocalDate, b: LocalDate) => expect(a.equals(b)).toBeTruthy()}
-    ${'Duration'}         | ${google.protobuf.Duration}                         | ${DurationGenerator}         | ${toDurationDao}  | ${fromDurationDao}         | ${(a: Duration, b: Duration) => expect(a.equals(b)).toBeTruthy()}
-    ${'TimeOnly'}         | ${Models.TimeOnlyDao}                               | ${LocalTimeGenerator}        | ${toTimeOnlyDao}  | ${fromTimeOnlyDao}         | ${(a: LocalTime, b: LocalTime) => expect(a.equals(b)).toBeTruthy()}
-    ${'SessionBlueprint'} | ${Models.SessionBlueprintDao.SessionBlueprintDaoV2} | ${SessionBlueprintGenerator} | ${modelToDao}     | ${fromSessionBlueprintDao} | ${toPOJOEquals}
-    ${'Session'}          | ${Models.SessionHistoryDao.SessionDaoV2}            | ${SessionGenerator}          | ${modelToDao}     | ${fromSessionDao}          | ${toPOJOEquals}
-    ${'FeedIdentity'}     | ${Models.FeedIdentityDaoV1}                         | ${FeedIdentityGenerator}     | ${modelToDao}     | ${fromFeedIdentityDao}     | ${toPOJOEquals}
-    ${'FeedUser'}         | ${Models.FeedUserDaoV1}                             | ${FeedUserGenerator}         | ${modelToDao}     | ${fromFeedUserDao}         | ${toPOJOEquals}
-    ${'SessionFeedItem'}  | ${Models.FeedItemDaoV1}                             | ${SessionFeedItemGenerator}  | ${modelToDao}     | ${fromFeedItemDao}         | ${toPOJOEquals}
-    ${'Timestamp'}        | ${google.protobuf.Timestamp}                        | ${InstantGenerator}          | ${toTimestampDao} | ${fromTimestampDao}        | ${(a: Instant, b: Instant) => a.equals(b)}
+    name                  | protoType                                           | initialValueGenerator        | convertToDao      | convertFromDao              | assertEquals
+    ${'Decimal'}          | ${Models.DecimalValue}                              | ${BigNumberGenerator}        | ${toDecimalDao}   | ${fromDecimalDao}           | ${(a: BigNumber, b: BigNumber) => expect(a.isEqualTo(b)).toBeTruthy()}
+    ${'DateOnly'}         | ${Models.DateOnlyDao}                               | ${LocalDateGenerator}        | ${toDateOnlyDao}  | ${fromDateOnlyDao}          | ${(a: LocalDate, b: LocalDate) => expect(a.equals(b)).toBeTruthy()}
+    ${'Duration'}         | ${google.protobuf.Duration}                         | ${DurationGenerator}         | ${toDurationDao}  | ${fromDurationDao}          | ${(a: Duration, b: Duration) => expect(a.equals(b)).toBeTruthy()}
+    ${'TimeOnly'}         | ${Models.TimeOnlyDao}                               | ${LocalTimeGenerator}        | ${toTimeOnlyDao}  | ${fromTimeOnlyDao}          | ${(a: LocalTime, b: LocalTime) => expect(a.equals(b)).toBeTruthy()}
+    ${'SessionBlueprint'} | ${Models.SessionBlueprintDao.SessionBlueprintDaoV2} | ${SessionBlueprintGenerator} | ${modelToDao}     | ${SessionBlueprint.fromDao} | ${toPOJOEquals}
+    ${'Session'}          | ${Models.SessionHistoryDao.SessionDaoV2}            | ${SessionGenerator}          | ${modelToDao}     | ${Session.fromDao}          | ${toPOJOEquals}
+    ${'FeedIdentity'}     | ${Models.FeedIdentityDaoV1}                         | ${FeedIdentityGenerator}     | ${modelToDao}     | ${FeedIdentity.fromDao}     | ${toPOJOEquals}
+    ${'FeedUser'}         | ${Models.FeedUserDaoV1}                             | ${FeedUserGenerator}         | ${modelToDao}     | ${FeedUser.fromDao}         | ${toPOJOEquals}
+    ${'SessionFeedItem'}  | ${Models.FeedItemDaoV1}                             | ${SessionFeedItemGenerator}  | ${modelToDao}     | ${SessionFeedItem.fromDao}  | ${toPOJOEquals}
+    ${'Timestamp'}        | ${google.protobuf.Timestamp}                        | ${InstantGenerator}          | ${toTimestampDao} | ${fromTimestampDao}         | ${(a: Instant, b: Instant) => a.equals(b)}
   `(
     'should convert back and forth between $name surviving an encoding',
     ({
