+++ app/store/feed/effects.ts
@@ -2,7 +2,6 @@ import { addEffect, getState } from '@/store/store';
 import {
   createFeedIdentity,
   feedApiError,
-  FeedState,
   fetchInboxItems,
   initializeFeedStateSlice,
   patchFeedState,
@@ -14,7 +13,6 @@ import {
   updateFeedIdentity,
 } from '@/store/feed';
 import { LiftLog } from '@/gen/proto';
-import { fromFeedStateDao } from '@/models/storage/conversions.from-dao';
 import { RemoteData } from '@/models/remote';
 import { addSharedItemEffects } from '@/store/feed/shared-item-effects';
 import { showSnackbar } from '@/store/app';
@@ -24,13 +22,7 @@ import { addFollowingEffects } from '@/store/feed/following-effects';
 import { selectActiveProgram } from '@/store/program';
 import { ApiErrorType, ApiResult } from '@/services/api-error';
 import { Platform } from 'react-native';
-import {
-  FeedIdentity,
-  FeedItem,
-  FeedUser,
-  FollowRequest,
-} from '@/models/feed-models';
-import { toUuidDao } from '@/models/storage/conversions.to-dao';
+import { fromFeedStateDao, toFeedStateDao } from './conversions';
 
 const StorageKey = 'FeedState';
 export function applyFeedEffects() {
@@ -325,27 +317,4 @@ export function applyFeedEffects() {
   addFeedItemEffects();
   addInboxEffects();
   addFollowingEffects();
-}
-
-export function toFeedStateDao(
-  state: FeedState,
-): LiftLog.Ui.Models.FeedStateDaoV1 {
-  return new LiftLog.Ui.Models.FeedStateDaoV1({
-    feedItems: state.feed.map((x) => FeedItem.fromPOJO(x).toDao()),
-    followedUsers: Object.values(state.followedUsers).map((x) =>
-      FeedUser.fromPOJO(x).toDao(),
-    ),
-    followers: Object.values(state.followers).map((x) =>
-      FeedUser.fromPOJO(x).toDao(),
-    ),
-    followRequests: state.followRequests.map((x) =>
-      FollowRequest.fromPOJO(x).toDao(),
-    ),
-    identity: state.identity
-      .map(FeedIdentity.fromPOJO)
-      .map((x) => x.toDao())
-      .unwrapOr(null),
-    unpublishedSessionIds: state.unpublishedSessionIds.map(toUuidDao),
-    revokedFollowSecrets: state.revokedFollowSecrets,
-  });
 }
